+++ lib/src/arculus_apple_button.dart
@@ -1,4 +1,5 @@
 import 'package:arculus_auth_widgets/src/apple_icon.dart';
+import 'package:arculus_auth_widgets/src/base_arculus_button.dart';
 import 'package:flutter/material.dart';
 
 /// Apple Sign In Button in Arculus Style which I think is neater.
@@ -13,12 +14,20 @@ class ArculusAppleButton extends StatelessWidget {
   /// Text to display on the button.
   final String label;
 
-  /// If null, the button will be displayed like a "disabled" buttons.
+  /// If null, the button will be displayed like a "disabled" button.
   final void Function(BuildContext) onPressed;
+
+  /// Displays [CircularProgressIndicator] at the center of the button.
+  /// The color of the indicator equals to [Colors.white] in light, and
+  /// [Colors.black] in dark.
+  ///
+  /// If true, will automatically disables [onPressed].
+  final bool isLoading;
 
   const ArculusAppleButton({
     Key key,
     @required this.label,
+    this.isLoading = false,
     this.onPressed,
   }) : super(key: key);
 
@@ -52,31 +61,38 @@ class ArculusAppleButton extends StatelessWidget {
     return baseColor;
   }
 
-  EdgeInsetsGeometry _getPadding(Set<MaterialState> states) {
+  EdgeInsetsGeometry _getPadding(
+    Set<MaterialState> states,
+    Brightness brightness,
+  ) {
     return EdgeInsets.fromLTRB(16, 16, 16, 16);
   }
 
   @override
   Widget build(BuildContext context) {
-    return ElevatedButton(
-      key: Key('arculus-apple-button'),
-      style: ButtonStyle(
-        foregroundColor: MaterialStateProperty.resolveWith(
-          (s) => _getForegroundColor(s, Theme.of(context).brightness),
-        ),
-        backgroundColor: MaterialStateProperty.resolveWith(
-          (s) => _getBackgroundColor(s, Theme.of(context).brightness),
-        ),
-        padding: MaterialStateProperty.resolveWith(_getPadding),
+    return Theme(
+      data: Theme.of(context).copyWith(
+        accentColor: Theme.of(context).brightness == Brightness.light
+            ? Colors.white.withOpacity(0.5)
+            : Colors.black.withOpacity(0.5),
       ),
-      child: Row(
-        children: [
-          SizedBox(child: AppleIcon(), width: 18, height: 18),
-          SizedBox(width: 32),
-          Text(label),
-        ],
+      child: BaseArculusButton(
+        key: Key('arculus-apple-button'),
+        isLoading: isLoading,
+        child: Container(
+          margin: const EdgeInsets.only(right: 16),
+          child: AppleIcon(
+            isEnabled: onPressed != null && !isLoading,
+          ),
+          width: 18,
+          height: 18,
+        ),
+        label: label,
+        onPressed: onPressed,
+        getForegroundColor: _getForegroundColor,
+        getBackgroundColor: _getBackgroundColor,
+        getPadding: _getPadding,
       ),
-      onPressed: onPressed != null ? () => onPressed(context) : null,
     );
   }
 }