+++ lib/src/arculus_google_button.dart
@@ -1,3 +1,4 @@
+import 'package:arculus_auth_widgets/src/base_arculus_button.dart';
 import 'package:arculus_auth_widgets/src/google_icon.dart';
 import 'package:flutter/material.dart';
 
@@ -10,15 +11,23 @@ import 'package:flutter/material.dart';
 ///
 /// See readme for more info.
 class ArculusGoogleButton extends StatelessWidget {
-  /// Text of the button
+  /// Text to display on the button.
   final String label;
 
-  /// If null, the button will be displayed like a "disabled" buttons.
+  /// If null, the button will be displayed like a "disabled" button.
   final void Function(BuildContext) onPressed;
+
+  /// Displays [CircularProgressIndicator] at the center of the button.
+  /// The color of the indicator equals to [Color(0xFF4285F4)] in light, and
+  /// [Theme.of(context).primaryTextTheme.button.color] in dark.
+  ///
+  /// If true, will automatically disables [onPressed].
+  final bool isLoading;
 
   const ArculusGoogleButton({
     Key key,
     @required this.label,
+    this.isLoading = false,
     this.onPressed,
   }) : super(key: key);
 
@@ -52,19 +61,22 @@ class ArculusGoogleButton extends StatelessWidget {
     return baseColor;
   }
 
-  EdgeInsetsGeometry _getDarkPadding(
-    Set<MaterialState> states, [
-    bool useArculusStyle = false,
-  ]) {
-    return EdgeInsets.fromLTRB(2, 2, 16, 2);
-  }
-
-  EdgeInsetsGeometry _getPadding(Set<MaterialState> states) {
+  EdgeInsetsGeometry _getPadding(
+      Set<MaterialState> states, Brightness brightness) {
+    if (brightness == Brightness.dark) {
+      return EdgeInsets.fromLTRB(0, 0, 16, 0);
+    }
     return EdgeInsets.fromLTRB(16, 16, 16, 16);
   }
 
   @override
   Widget build(BuildContext context) {
+    Widget icon = SizedBox(
+      width: 18,
+      height: 18,
+      child: GoogleIcon(),
+    );
+
     if (Theme.of(context).brightness == Brightness.dark) {
       OutlinedBorder _parentBorder = Theme.of(context)
           .elevatedButtonTheme
@@ -82,57 +94,37 @@ class ArculusGoogleButton extends StatelessWidget {
         _borderRadius = _parentBorder.borderRadius;
       }
 
-      return ElevatedButton(
-        key: Key('arculus-google-button'),
-        style: ButtonStyle(
-          foregroundColor: MaterialStateProperty.resolveWith(
-            (s) => _getForegroundColor(s, Theme.of(context).brightness),
-          ),
-          backgroundColor: MaterialStateProperty.resolveWith(
-            (s) => _getBackgroundColor(s, Theme.of(context).brightness),
-          ),
-          padding: MaterialStateProperty.resolveWith(_getDarkPadding),
+      icon = Container(
+        decoration: BoxDecoration(
+          borderRadius: _borderRadius,
+          color: Colors.white
+              .withOpacity(onPressed != null && !isLoading ? 1 : 0.5),
         ),
-        child: Row(
-          children: [
-            Container(
-              decoration: BoxDecoration(
-                borderRadius: _borderRadius,
-                color: Colors.white,
-              ),
-              padding: const EdgeInsets.all(14),
-              child: SizedBox(width: 18, height: 18, child: GoogleIcon()),
-            ),
-            SizedBox(width: 16),
-            Text(label),
-          ],
-        ),
-        onPressed: onPressed != null ? () => onPressed(context) : null,
+        margin: const EdgeInsets.all(2),
+        padding: const EdgeInsets.all(14),
+        child: SizedBox(
+            width: 18,
+            height: 18,
+            child: GoogleIcon(isEnabled: onPressed != null && !isLoading)),
       );
     }
 
-    return ElevatedButton(
-      style: ButtonStyle(
-        foregroundColor: MaterialStateProperty.resolveWith(
-          (s) => _getForegroundColor(s, Theme.of(context).brightness),
-        ),
-        backgroundColor: MaterialStateProperty.resolveWith(
-          (s) => _getBackgroundColor(s, Theme.of(context).brightness),
-        ),
-        padding: MaterialStateProperty.resolveWith(_getPadding),
+    return Theme(
+      data: Theme.of(context).copyWith(
+        accentColor: Theme.of(context).brightness == Brightness.light
+            ? Color(0xFF4285F4)
+            : Theme.of(context).primaryTextTheme.button.color.withOpacity(0.5),
       ),
-      child: Row(
-        children: [
-          SizedBox(
-            width: 18,
-            height: 18,
-            child: GoogleIcon(),
-          ),
-          SizedBox(width: 32),
-          Text(label),
-        ],
+      child: BaseArculusButton(
+        key: Key('arculus-google-button'),
+        isLoading: isLoading,
+        child: icon,
+        label: label,
+        onPressed: onPressed,
+        getForegroundColor: _getForegroundColor,
+        getBackgroundColor: _getBackgroundColor,
+        getPadding: _getPadding,
       ),
-      onPressed: onPressed != null ? () => onPressed(context) : null,
     );
   }
 }